refactor(codegen,core,shared): consolidate extractor package into shared and reorganize utilities
- Rename @mearie/extractor to @mearie/shared for better semantic clarity
- Move extraction logic (loaders, parsers, extractor) from extractor to codegen
- Move shared utilities (errors, logger) from core to codegen
- Extract common types from core to shared package
- Update import paths across affected packages (core, codegen, vite)
- Remove extractor package artifacts (CHANGELOG, README)
