+++ .changeset/pre.json
@@ -8,7 +8,6 @@
     "@mearie/codegen": "0.0.0",
     "@mearie/config": "0.0.0",
     "@mearie/core": "0.0.0",
-    "@mearie/extractor": "0.0.0",
     "mearie": "0.0.0",
     "@mearie/react": "0.0.0",
     "@mearie/solid": "0.0.0",
