+++ .changeset/silent-years-jump.md
@@ -1,0 +1,8 @@
+---
+"@mearie/codegen": minor
+"@mearie/core": minor
+"@mearie/shared": minor
+"@mearie/vite": patch
+---
+
+Consolidate extractor package into shared package and move internal utilities to codegen
