+++ packages/codegen/package.json
@@ -17,17 +17,21 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "@logtape/logtape": "^1.1.1",
     "@mearie/config": "workspace:*",
-    "@mearie/core": "workspace:*",
-    "@mearie/extractor": "workspace:*",
     "@mearie/native": "workspace:*",
+    "@mearie/shared": "workspace:*",
+    "picocolors": "^1.1.1",
     "picomatch": "^4.0.3",
     "tinyglobby": "^0.2.15"
   },
   "devDependencies": {
     "@types/node": "catalog:",
     "@types/picomatch": "^4.0.2",
+    "@vue/compiler-sfc": "^3.5.22",
+    "svelte": "^5.41.0",
     "tsdown": "catalog:",
+    "typescript": "^5.9.3",
     "vitest": "catalog:"
   },
   "publishConfig": {
