+++ packages/codegen/src/index.ts
@@ -1,4 +1,3 @@
-export { findFiles, createMatcher, type GlobPatterns } from './glob.ts';
-export { generate, type GenerateOptions } from './generator.ts';
+export { findFiles, createMatcher } from './glob.ts';
 export { CodegenContext } from './context.ts';
-export type { CodegenConfig } from './types.ts';
+export { logger, report } from './logger.ts';
