+++ packages/codegen/src/types.ts
@@ -1,9 +1,5 @@
-/**
- * Configuration for the code generator.
- */
-export type CodegenConfig = {
-  schemas: string;
-  documents: string | string[];
-  exclude?: string | string[];
-  cwd: string;
+export type Source = {
+  code: string;
+  filePath: string;
+  startLine: number;
 };
