+++ packages/core/src/types.ts
@@ -1,45 +1,4 @@
-export type MaybePromise<T> = T | Promise<T>;
-
-export type Nullable<T> = T | null;
-
-export type List<T> = readonly T[];
-
-export type Opaque<T> = T & { readonly ' $opaque'?: unique symbol };
-
-export type Source = {
-  code: string;
-  filePath: string;
-  startLine: number;
-};
-
-export type ArtifactKind = 'query' | 'mutation' | 'subscription' | 'fragment';
-
-export type Artifact<
-  Kind extends ArtifactKind = ArtifactKind,
-  Name extends string = string,
-  Data = unknown,
-  Variables = unknown,
-> = {
-  readonly kind: Kind;
-  readonly name: Name;
-  readonly source: string;
-  readonly selections: readonly Selection[];
-
-  readonly ' $data'?: Data;
-  readonly ' $variables'?: Variables;
-};
-
-export type Selection = {
-  name: string;
-  type?: string;
-  array?: boolean;
-  on?: string[];
-  alias?: string;
-  args?: Record<string, Argument>;
-  selections?: Selection[];
-};
-
-export type Argument = { kind: 'literal'; value: unknown } | { kind: 'variable'; name: string };
+import type { Artifact } from '@mearie/shared';
 
 export type SchemaMeta = {
   entities: Record<string, EntityMeta>;
@@ -55,10 +14,3 @@ export type Operation<T extends Artifact<'query' | 'mutation' | 'subscription'>>
   variables?: T[' $variables'];
   signal?: AbortSignal;
 };
-
-export type FragmentRefs<T extends string> = {
-  readonly ' $fragmentRefs': T;
-};
-
-export type DataOf<T extends Artifact> = NonNullable<T[' $data']>;
-export type VariablesOf<T extends Artifact> = NonNullable<T[' $variables']>;
