--- packages/extractor/CHANGELOG.md
@@ -1,37 +1,0 @@
-# @mearie/extractor
-
-## 0.0.1-next.3
-
-### Patch Changes
-
-- 45b907b: Fix native package publish pipeline
-- Updated dependencies [45b907b]
-  - @mearie/native@0.0.1-next.3
-  - @mearie/core@0.0.1-next.3
-
-## 0.0.1-next.2
-
-### Patch Changes
-
-- 80fdb0c: Fix release pipeline
-- Updated dependencies [80fdb0c]
-  - @mearie/native@0.0.1-next.2
-  - @mearie/core@0.0.1-next.2
-
-## 0.0.1-next.1
-
-### Patch Changes
-
-- 7dd09dd: chore(release): version packages
-- Updated dependencies [7dd09dd]
-  - @mearie/native@0.0.1-next.1
-  - @mearie/core@0.0.1-next.1
-
-## 0.0.1-next.0
-
-### Patch Changes
-
-- 0aa1561: chore(release): version packages
-- Updated dependencies [0aa1561]
-  - @mearie/native@0.0.1-next.0
-  - @mearie/core@0.0.1-next.0
