--- packages/extractor/README.md
@@ -1,11 +1,0 @@
-# @mearie/extractor
-
-GraphQL query extractor for Mearie.
-
-This package extracts GraphQL operations from source files (JavaScript,
-TypeScript, Vue, Svelte). It is used internally by build plugins and is not
-intended for direct use.
-
-## Documentation
-
-Full documentation is available at <https://mearie.dev/>.
