+++ packages/shared/README.md
@@ -1,0 +1,5 @@
+# @mearie/shared
+
+Common types and error classes shared across all Mearie packages.
+
+This package contains only TypeScript types and error class definitions, with zero runtime dependencies.
