+++ packages/shared/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mearie/extractor",
+  "name": "@mearie/shared",
   "version": "0.0.1-next.3",
   "sideEffects": false,
   "type": "module",
@@ -16,16 +16,8 @@
     "test": "vitest run",
     "test:watch": "vitest"
   },
-  "dependencies": {
-    "@mearie/core": "workspace:*",
-    "@mearie/native": "workspace:*"
-  },
   "devDependencies": {
-    "@types/babel__traverse": "^7.28.0",
-    "@vue/compiler-sfc": "^3.5.22",
-    "svelte": "^5.41.0",
     "tsdown": "catalog:",
-    "typescript": "^5.9.3",
     "vitest": "catalog:"
   },
   "publishConfig": {
