+++ packages/shared/tsdown.config.ts
@@ -5,5 +5,4 @@ export default defineConfig({
   format: ['cjs', 'esm'],
   clean: true,
   dts: true,
-  external: ['@mearie/native', '@vue/compiler-sfc', 'svelte', 'svelte/compiler', 'svelte2tsx', 'typescript'],
 });
